Revert "docs: remove superfluous SITE_URL from .env.local"
This reverts commit 396b91c9b5b5e9a86d5bf852a88cc808bc15a0d2.

Full stack framework headers require the baseURL to be defined the
same as it is in Convex because the url being http or https affects
whether Better Auth uses secure cookies. A mismatch means the framework
server may end up looking for an insecure cookie when a secure cookie
has been set, or vice versa, leading to a broken implementation.